Refactorización y mejora de la capa de persistencia para la gestión de productos
- Se agregó la anotación @Repository a ProductoRepository para escanear componentes de Spring.
- Se introdujeron nuevos métodos en ProductoRepository para obtener productos por categoría y compras de clientes.
- Se actualizó ProductoCrudRepository para incluir un nuevo método para filtrar productos por stock y estado.
- Se anotaron las clases Categoria y Cliente con @Entity y @Table para mapeo JPA.
- Se cambiaron los tipos de datos en las clases ComprasProducto y Producto de Double a BigDecimal para una mejor precisión en los cálculos financieros.
- Se refactorizaron los nombres de los campos en Producto para lograr coherencia y claridad.

+++ src/main/java/com/crislearning/soncco_market/persistence/entity/ComprasProducto.java
@@ -1,5 +1,7 @@
 package com.crislearning.soncco_market.persistence.entity;
 
+import java.math.BigDecimal;
+
 import jakarta.persistence.Column;
 import jakarta.persistence.EmbeddedId;
 import jakarta.persistence.Entity;
@@ -18,7 +20,7 @@ public class ComprasProducto {
     private Integer cantidad;
 
     @Column(name = "total", precision = 16, scale = 2)
-    private Double total;
+    private BigDecimal total;
 
     @Column(name = "estado")
     private Integer estado;
@@ -47,11 +49,11 @@ public class ComprasProducto {
         this.cantidad = cantidad;
     }
 
-    public Double getTotal() {
+    public BigDecimal  getTotal() {
         return total;
     }
 
-    public void setTotal(Double total) {
+    public void setTotal(BigDecimal  total) {
         this.total = total;
     }
